DHMx-c: Changed pyqtSignal's timestamp from int to a float

+++ dhm_gui/display_dhmxc.py
@@ -34,7 +34,7 @@ headerStruct = struct.Struct('QQQQQQQdddddddd')
 class guiclient(QThread):
     sig_img = pyqtSignal(int,str, np.ndarray)
     sig_hist_val = pyqtSignal(int,int)
-    sig_header = pyqtSignal(int, int, int, int, int, int, int, int) #header: height, width, frameID, timestamp, gain min, gain max, exposure min, exposure max
+    sig_header = pyqtSignal(int, int, int, float, int, int, int, int) #header: height, width, frameID, timestamp, gain min, gain max, exposure min, exposure max
     sig_img_complete = pyqtSignal(str)
     def __init__(self,server,port):
         QThread.__init__(self)
